Refactor setup to take extra setup functions ⛽
And use this to extract the "local" registry part of TestKanikoTaskRun
into a separate function, reusable in other tests.
